pyproject.toml
[project]
name = "mini-grpo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.4.2",
    "instructlab-training>=0.12.1",
    "ipython>=9.8.0",
    "typer>=0.21.0",
]
